+++ src/main_loop/main_loop_console.c
@@ -50,8 +50,8 @@ static ret_t main_loop_console_recv_event_mutex(main_loop_t* l, event_queue_req_
 
 static ret_t main_loop_dispatch_events(main_loop_console_t* loop) {
   event_queue_req_t r;
-  int time_in = time_now_ms();
-  int time_out = time_in;
+  uint64_t time_in = time_now_ms();
+  uint64_t time_out = time_in;
 
   while ((time_out - time_in < 20) && (main_loop_recv_event((main_loop_t*)loop, &r) == RET_OK)) {
     switch (r.event.type) {
