+++ src/main_loop/main_loop_simple.c
@@ -173,8 +173,8 @@ ret_t main_loop_post_key_event(main_loop_t* l, bool_t pressed, uint8_t key) {
 
 static ret_t main_loop_dispatch_events(main_loop_simple_t* loop) {
   event_queue_req_t r;
-  int time_in = time_now_ms();
-  int time_out = time_in;
+  uint64_t time_in = time_now_ms();
+  uint64_t time_out = time_in;
 
   while ((time_out - time_in < 20) && (main_loop_recv_event((main_loop_t*)loop, &r) == RET_OK)) {
     widget_t* widget = loop->base.wm;
